Add flex properties, padding, and margin
For the header, I set a flex display, centered items vertically, added padding, and a flex factor. For the ul, I set a flex display, a flex factor, got rid of padding, and set a gap for the items. For the left side of the header, I gave it a flex factor, put a gap for the items, added margin, and had the content flush left. For the right side of the header, I did the same as the left except had flushed content to the right.