Wrap title and items into their own divs
I wrapped the title into a div to easily be able to center the title. I wrapped all the images and texts into a div to be able to align all of the items on the same line. I wrapped each image and corresponding text into its own div to then be able to have the image and text be able to effectively set the width of each item, as well as being able to have the text wrap.