Add flex properties to title and items
For the title, I made sure to center it on the page as well as setting a margin on the bottom to give space between the title and the items. For the items, I made sure to have the items centered, as well as giving a gap between each item. For each individual item, I made sure to have the text wrap onto its own line.