Add flex properties, padding, and font-weight
I gave padding to the modal, and set the display to flex. For the icon and close button, I set a flex shink factor of 0 to prevent these pieces from being squished. For the text and buttons div, I gave it a flex direction of column and also a gap between the items. Finally, I made the header bold and a larger font size.