Make content centered, style header and footer
For the header and footer, I made sure to keep these items at the top and bottom, respectively, by setting the body to column flex. For both the header and footer, I pushed the links to either side of the header/footer, and styled the links to not have the standardized blue color and text underline. Finally, for the main content, I set this piece to flex column, and wrapped the content onto their own lines.