Centered main content, styled header and footer
For the body, I applied flex column with space-between so that the footer and header were flushed on the edges. I gave the header full width, and centered the text. For the sidebar, I set a column flex and made sure to also give the sidebar a flex shrink and grow factor of 0 to maintain a 300px width. For the card section, I defined the height and width for each card, and gave the whole card section a flex wrap so that cards could continue onto next line(s). Finally, I set font sizing, removed text decoration from anchor elements, and removed bullets from li elements.